separate endpoint from models, add app_probe list
