+++ catalystwan/endpoints/configuration/policy/list/app_probe.py
@@ -4,23 +4,14 @@
 from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
-from catalystwan.models.policy.lists import AppProbeClassList
-from catalystwan.models.policy.policy_list import (
-    InfoTag,
-    PolicyListEndpoints,
-    PolicyListId,
-    PolicyListInfo,
-    PolicyListPreview,
+from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
+from catalystwan.models.policy.list.app_probe import (
+    AppProbeClassList,
+    AppProbeClassListEditPayload,
+    AppProbeClassListInfo,
 )
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class AppProbeClassListEditPayload(AppProbeClassList, PolicyListId):
-    pass
-
-
-class AppProbeClassListInfo(AppProbeClassList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyAppProbeClassList(APIEndpoints, PolicyListEndpoints):
