+++ catalystwan/endpoints/configuration/policy/list/ips_signature.py
@@ -4,14 +4,9 @@
 from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
+from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
 from catalystwan.models.policy.lists import IPSSignatureList
-from catalystwan.models.policy.policy_list import (
-    InfoTag,
-    PolicyListEndpoints,
-    PolicyListId,
-    PolicyListInfo,
-    PolicyListPreview,
-)
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
 from catalystwan.typed_list import DataSequence
